refactor(store): break import cycles and isolate cross-module orchestrator
`npm run lint:arch` failed with 7 circular store imports plus a renderer-
imports-main violation. Untangled without behavior changes:

- `src/store/focused-panel.ts` (new): the focus-registry getters/setters
  (`triggerFocus`, `getTaskFocusedPanel`, `setTaskFocusedPanel`,
  `isPanelFocused`, `defaultPanelFor`, `aiTerminalPanels`, etc.) live
  here. `focus.ts` keeps grid navigation and re-exports the moved
  symbols for backward compatibility with existing call sites.
- `src/store/project-cleanup.ts` (new): `removeProjectWithTasks` is now
  the lone orchestrator that depends on both projects and tasks, so
  `projects.ts` no longer needs to import `closeTask`.
- Updated `navigation`, `tasks`, `terminals`, `ui` (and `ui.test.ts`
  mock target) to import focused-panel helpers from the new module
  instead of round-tripping through `focus.ts`.

After this change the projects↔tasks, persistence chain, navigation↔
focus, and focus↔tasks cycles are gone.

- .dependency-cruiser.cjs: `electron/mcp/prompt-detect.ts` is a pure
  regex module already shared with the renderer task-status pipeline;
  whitelist it next to `electron/ipc/channels.ts` so the rule's intent
  (block Node/Electron-bound code) remains while the documented shared
  module is allowed.

Addresses finding 2 of issue #157.
